assignment 3
