+++ life_expectancy/cleaning.py
@@ -1,20 +1,7 @@
-import argparse
 from pathlib import Path
 
 import pandas as pd
 
-
-def load_data() -> pd.DataFrame:
-    """
-    Method used to load data.
-
-    Returns:
-        [pd.DataFrame]: Dataframe to be cleaned.
-    """
-    data_df = pd.read_csv(
-        f"{Path(__file__).parent}/data/eu_life_expectancy_raw.tsv", sep='\t')
-
-    return data_df
 
 def clean_data(data_df: pd.DataFrame, region: str = 'PT') -> pd.DataFrame:
     """
@@ -59,31 +46,3 @@ def clean_data(data_df: pd.DataFrame, region: str = 'PT') -> pd.DataFrame:
     clean_df = clean_df.dropna(subset=['value'])
 
     return clean_df
-
-
-def save_data(clean_df: pd.DataFrame) -> None:
-    """
-    Method used to save data.
-    Args:
-        clean_df [pd.DataFrame]: Cleaned and filtered DataFrame to be saved.
-    """
-
-    # Save the resulting data frame to the data folder as pt_life_expectancy.csv
-    clean_df.to_csv(
-            f"{Path(__file__).parent}/data/pt_life_expectancy.csv", index = False)
-
-
-def main(region: str) -> None:
-    """
-    Main function.
-    Method that loads, cleans and saves data.
-    """
-    data_df = load_data()
-    clean_df = clean_data(data_df, region)
-    save_data(clean_df)
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument('region')
-    args = parser.parse_args()
-    main(args.region)
